solidity 0.7.6: Convert bytes32 to address
See https://docs.soliditylang.org/en/v0.7.6/types.html#address

+++ contracts/DogeParser/DogeMessageLibrary.sol
@@ -692,8 +692,8 @@ library DogeMessageLibrary {
           y = p - y;
         }
         require(yy == mulmod(y, y, p));
-        // Now, with uncompressed x and y, create the address
-        return address(keccak256(abi.encodePacked(x, y)));
+        // Now, with uncompressed x and y, create the address         
+        return address(uint160(uint256(keccak256(abi.encodePacked(x, y)))));
     }
 
     // Gets the public key hash given a public key
